cleanup FlightShufflePartitionRef
